editing comments

+++ app/Livewire/CommentItem.php
@@ -9,6 +9,12 @@ class CommentItem extends Component
 {
     public Comment $comment;
 
+    public bool $editing = false;
+    protected $listeners = [
+        'cancelEditing' => 'cancelEditing',
+        'commentUpdated' => 'commentUpdated',
+    ];
+
     public function mount(Comment $comment)
     {
         $this->comment = $comment;
@@ -21,9 +27,33 @@ class CommentItem extends Component
 
     public function deleteComment()
     {
+        $user = auth()->user();
+        if(!$user) {
+            return $this->redirect('/login');
+        }
+
+        if($this->comment->user_id != $user->id) {
+            return response('Not allowed to edit comment', 403);
+        }
+
         $id = $this->comment->id;
         $this->comment->delete();
 
         $this->dispatch('commentDeleted', $id);
+    }
+
+    public function startCommentEdit()
+    {
+        $this->editing = true;
+    }
+
+    public function cancelEditing()
+    {
+        $this->editing = false;
+    }
+
+    public function commentUpdated()
+    {
+        $this->editing = false;
     }
 }
